requirements.txt

tkinter
pillow
dlib
scikit-image
